(split)http://bugzilla.zimbra.com/show_bug.cgi?id=6015 bug: 6015 use millisecond resolution timestamp for consistency. add fragment. return creator and last modified by attributes on search and getwiki request.
Copied from Perforce
 Change: 23113

+++ src/java/com/zimbra/cs/wiki/WikiWord.java
@@ -37,7 +37,7 @@ public class WikiWord {
 	private Mailbox mMailbox;
 	private int mId;
 	private int mRevision;
-	private int mModifiedDate;
+	private long mModifiedDate;
 	private String mModifiedBy;
 	private String mCreator;
 	private int mFolderId;
